Fix cycle dependecy issues
